+++ src/tools/drill.rs
@@ -10,12 +10,12 @@ impl Drill {
         id: u32,
         name: String,
         note: String,
-        shaft_diameter: f32,
-        tool_diameter: f32,
-        spindle_speed: f32,
-        pass_depth: f32,
-        plunge_rate: f32,
-        feed_rate: f32,
+        shaft_diameter: f64,
+        tool_diameter: f64,
+        spindle_speed: f64,
+        pass_depth: f64,
+        plunge_rate: f64,
+        feed_rate: f64,
     ) -> Self {
         Self {
             base_tool: BaseTool::new(
@@ -38,27 +38,27 @@ impl Tool for Drill {
     //     self.base_tool.name.clone()
     // }
 
-    fn offset(&self, _depth_cut: f32) -> f32 {
+    fn offset(&self, _depth_cut: f64) -> f64 {
         self.base_tool.tool_diameter / 2.0
     }
 
-    // fn shaft_diameter(&self) -> f32 {
+    // fn shaft_diameter(&self) -> f64 {
     //     self.base_tool.shaft_diameter
     // }
 
-    // fn single_pass_depth(&self) -> f32 {
+    // fn single_pass_depth(&self) -> f64 {
     //     self.base_tool.single_pass_depth
     // }
 
-    // fn spindle_speed(&self) -> f32 {
+    // fn spindle_speed(&self) -> f64 {
     //     self.base_tool.spindle_speed
     // }
 
-    // fn vertical_speed(&self) -> f32 {
+    // fn vertical_speed(&self) -> f64 {
     //     self.base_tool.vertical_speed
     // }
 
-    // fn horizontal_speed(&self) -> f32 {
+    // fn horizontal_speed(&self) -> f64 {
     //     self.base_tool.horizontal_speed
     // }
 }
